+++ client/src/components/graph/graph.js
@@ -6,8 +6,7 @@ import { connect } from "react-redux";
 import mat4 from "gl-mat4";
 import _regl from "regl";
 
-import { FaCrosshairs } from "react-icons/fa";
-import { FaSearchPlus } from "react-icons/fa";
+import { FaCrosshairs, FaSearchPlus } from "react-icons/fa";
 
 import * as globals from "../../globals";
 import setupSVGandBrushElements from "./setupSVGandBrush";
