+++ client/src/components/scatterplot/setupScatterplot.js
@@ -8,14 +8,14 @@
 import * as d3 from "d3";
 
 const setupScatterplot = (width, height, margin) => {
-  var container = d3.select("#scatterplot");
+  const container = d3.select("#scatterplot");
 
-  var svg = container
+  const svg = container
     .append("svg")
     .attr("width", width + margin.left + margin.right)
     .attr("height", height + margin.top + margin.bottom)
     .append("g")
-    .attr("transform", "translate(" + margin.left + "," + margin.top + ")");
+    .attr("transform", `translate(${margin.left},${margin.top})`);
 
   return {
     svg
