+++ client/src/components/scatterplot/util.js
@@ -1,9 +1,4 @@
 // jshint esversion: 6
-import _ from "lodash";
-
-const paddingRight = 120;
-const continuousChartWidth = 340;
-
 export const margin = { top: 66, right: 110, bottom: 20, left: 60 };
 export const width = 340;
 export const height = 340 - margin.top - margin.bottom;
